fix: convert simple-password-provider to typescript
